+++ client/src/app/app.module.ts
@@ -70,7 +70,7 @@ import {
 } from '@angular-material-components/datetime-picker';
 import { StepsService } from './services/steps.service';
 import { GenericListComponent } from 'src/app/components/generic/list/generic-list.component';
-import { ThemeService } from './services/theme.service.ts.service';
+import { ThemeService } from './services/theme.service';
 
 @NgModule({
   declarations: [
